refactor: package 세분화
