+++ backend/src/main/java/team/teamby/teambyteam/sse/domain/converter/TeamPlaceSseConverter.java
@@ -1,6 +1,7 @@
-package team.teamby.teambyteam.sse.domain;
+package team.teamby.teambyteam.sse.domain.converter;
 
 import team.teamby.teambyteam.common.domain.DomainEvent;
+import team.teamby.teambyteam.sse.domain.TeamPlaceSseEvent;
 
 public interface TeamPlaceSseConverter {
     TeamPlaceSseEvent convert(DomainEvent event);
