+++ backend/src/main/java/team/teamby/teambyteam/sse/domain/emitter/SseEmitters.java
@@ -1,9 +1,11 @@
-package team.teamby.teambyteam.sse.domain;
+package team.teamby.teambyteam.sse.domain.emitter;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
+import team.teamby.teambyteam.sse.domain.TeamPlaceEventId;
+import team.teamby.teambyteam.sse.domain.TeamPlaceSseEvent;
 
 import java.io.IOException;
 import java.util.Map;
