+++ backend/src/main/java/team/teamby/teambyteam/sse/domain/emitter/TeamPlaceEmitterRepository.java
@@ -1,4 +1,4 @@
-package team.teamby.teambyteam.sse.domain;
+package team.teamby.teambyteam.sse.domain.emitter;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
